Refactoring, styling
