+++ public/styles/slider.scss
