+++ public/styles/style.scss
@@ -15,130 +15,6 @@
 
 $primary-color: #47a7ff;
 
-html,
-body,
-div,
-span,
-applet,
-object,
-iframe,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-blockquote,
-pre,
-a,
-abbr,
-acronym,
-address,
-big,
-cite,
-code,
-del,
-dfn,
-em,
-img,
-ins,
-kbd,
-q,
-s,
-samp,
-small,
-strike,
-sub,
-sup,
-tt,
-var,
-b,
-u,
-i,
-center,
-dl,
-dt,
-dd,
-ol,
-ul,
-li,
-fieldset,
-form,
-label,
-legend,
-table,
-caption,
-tbody,
-tfoot,
-thead,
-tr,
-th,
-td,
-article,
-aside,
-canvas,
-details,
-embed,
-figure,
-figcaption,
-footer,
-header,
-hgroup,
-menu,
-nav,
-output,
-ruby,
-section,
-summary,
-time,
-mark,
-audio,
-video {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  font-size: 100%;
-  font: inherit;
-  vertical-align: baseline;
-}
-/* HTML5 display-role reset for older browsers */
-article,
-aside,
-details,
-figcaption,
-figure,
-footer,
-header,
-hgroup,
-menu,
-nav,
-section {
-  display: block;
-}
-body {
-  line-height: 1;
-}
-ol,
-ul {
-  list-style: none;
-}
-blockquote,
-q {
-  quotes: none;
-}
-blockquote:before,
-blockquote:after,
-q:before,
-q:after {
-  content: '';
-  content: none;
-}
-table {
-  border-collapse: collapse;
-  border-spacing: 0;
-}
-
 body,
 html {
   padding: 0;
@@ -996,16 +872,10 @@ header {
     height: calc(100% - 45px);
   }
 
-  &:hover {
-    transform: scale(1.02, 1.02);
-  }
-
   &.isBeingDragged,
   &.selected.isBeingDragged {
     border: 1.5px solid #47a7ff;
     box-sizing: border-box;
-    // box-shadow: 0px 5.60708px 16.8212px rgba(111, 123, 138, 0.2),
-    //   0px 1.40177px 2.80354px rgba(111, 123, 138, 0.2), 0px 0px 0px rgba(111, 123, 138, 0.1);
     background: linear-gradient(
         0deg,
         rgba(71, 167, 255, 0.15),
@@ -1766,5 +1636,44 @@ button,
 .bp4-light {
   .meta {
     color: $pt-text-color-muted;
+  }
+}
+
+// .x6-selecting {
+//   .selected {
+//   }
+// }
+
+.x6-widget-transform {
+  border: 2px solid #ffae64;
+  box-sizing: border-box !important;
+  margin: 0;
+  padding: 0;
+
+  div {
+    border: 2px solid #ffae64;
+  }
+
+  .x6-widget-transform-resize {
+    border-radius: 2px;
+    width: 14px;
+    height: 14px;
+
+    &[data-position='right'],
+    &[data-position='left'],
+    &[data-position='top'],
+    &[data-position='bottom'] {
+      display: none;
+    }
+  }
+}
+
+.x6-toolbar-content {
+  width: 100%;
+  height: 100%;
+  .dnd-wrap {
+    display: flex;
+    align-items: center;
+    justify-content: center;
   }
 }
