+++ public/styles/toastify/_closeButton.scss
