+++ public/styles/toastify/_progressBar.scss
