+++ public/styles/toastify/animations/_slide.scss
