+++ public/styles/toastify/minimal.scss
